+++ test/DeferTest.js
@@ -1,17 +1,17 @@
 "use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
 /// <reference types="chai" />
 /// <reference types="mocha" />
-Object.defineProperty(exports, "__esModule", { value: true });
 var chai = require("chai");
 var Defer = require("../Defer");
 var asr = chai.assert;
 suite("Defer", function DeferTest() {
     test("when", function whenTest(done) {
-        var a = Defer.newPromiseResolved("start");
-        var b = Defer.newPromiseResolved({ s: 22 });
-        var c = Defer.newPromiseResolved(10);
+        var a = Defer.resolve("start");
+        var b = Defer.resolve({ s: 22 });
+        var c = Defer.resolve(10);
         var dfd = Defer.newDefer();
-        var p1 = dfd.promise.then(function () { return Defer.newPromiseResolved(null); }, function () { return Defer.newPromiseResolved({ prop: new String(23) }); });
+        var p1 = dfd.promise.then(function () { return Defer.resolve(null); }, function () { return Defer.resolve({ prop: new String(23) }); });
         var p2 = p1.then(function () { return ({ prop: 23 }); });
         var p3 = p2.then(function () { return c; }, function (err) { return (Math.random() > 0.5 ? err : Defer.throwBack({ boom: "error" })); });
         var p4 = p3.then(function (res) { return res; }, function (err) { return Defer.throwBack(c); });
@@ -24,10 +24,31 @@ suite("Defer", function DeferTest() {
             done();
         });
     });
+    test("then", function thenTest(done) {
+        var p = Defer.resolve({ prop: 23 });
+        p.then(function (r) {
+            if (r.prop > 10) {
+                return Defer.resolve(r.prop);
+            }
+            else {
+                return Defer.reject({ errText: "p-error" });
+            }
+        }).then(function (r2) {
+            return r2.toFixed();
+        }).catch(function (err) {
+            return Defer.throw(err.message ? { errText: err.message } : err);
+        }).then(function (r3) {
+            asr.equal(r3, "23");
+            done();
+        }, function (e3) {
+            asr.ok(false, "unexpecte error: " + JSON.stringify(e3));
+            done();
+        });
+    });
     test("throwBack", function throwBackTest(done) {
         try {
             var res = Defer.throwBack("my-error");
-            asr.equal(false, true, "expected error to be thrown by throwBack()");
+            asr.ok(false, "expected error to be thrown by throwBack()");
             done();
         }
         catch (err) {
@@ -35,21 +56,46 @@ suite("Defer", function DeferTest() {
             done();
         }
     });
+    test("catch simple", function catchSimpleTest(done) {
+        var init = Defer.reject({ err: "hmm" });
+        var r1 = init.catch(function (err) {
+            asr.equal(err.err, "hmm");
+        }).then(function (res) {
+            asr.equal(res, undefined);
+        }, function (err) {
+            asr.ok(false, "unexpected error: " + JSON.stringify(err));
+        });
+        var r2 = init.catch(function (err) {
+            // do nothing
+        });
+        Defer.when([r1, r2]).then(function (_a) {
+            var res1 = _a[0], res2 = _a[1];
+            asr.deepEqual(res1, undefined);
+            asr.deepEqual(res2, undefined);
+            done();
+        }, function (err) {
+            asr.isNotOk(err);
+            done();
+        });
+    });
     test("catch", function catchTest(done) {
-        var b = Defer.newPromiseRejected({ errText: "testing catch" });
-        var r1 = b.then(function (res) {
+        var init = Defer.reject({ errText: "testing catch" });
+        var r1 = init.then(function (res) {
             return { res2: "return result" };
         }, function (err) {
             return { err2: "return from error catch" };
         });
-        var r2 = b.then(function (res) {
-            return { res2: "return result 2" };
-        }, function (err) {
+        function rr(err) {
             return err ? Defer.throwBack({ errErr: "throw from error catch" }) : { err2: "return from error catch 2" };
-        }).catch(function (err) {
+        }
+        var rrr = init.catch(rr);
+        var r2 = init.then(function (res) {
+            return { res2: "return result 2" };
+        }, rr);
+        r2 = r2.catch(function (err) {
             return err;
         });
-        var r3 = b.catch(function (err) {
+        var r3 = init.catch(function (err) {
             if (err.errText) {
                 return err.errText;
             }
@@ -57,11 +103,15 @@ suite("Defer", function DeferTest() {
                 return Defer.throwBack(err.errText);
             }
         });
-        Defer.when([r1, r2, r3]).then(function (_a) {
-            var res1 = _a[0], res2 = _a[1], res3 = _a[2];
+        var r4 = init.catch(function (err) {
+            // return nothing
+        });
+        Defer.when([r1, r2, r3, r4]).then(function (_a) {
+            var res1 = _a[0], res2 = _a[1], res3 = _a[2], res4 = _a[3];
             asr.deepEqual(res1, { err2: "return from error catch" });
             asr.deepEqual(res2, { errErr: "throw from error catch" });
             asr.equal(res3, "testing catch");
+            asr.equal(res4, undefined);
             done();
         }, function (err) {
             asr.isNotOk(err);
@@ -89,7 +139,7 @@ suite("Defer", function DeferTest() {
     });
     test("runActionForAllInSeries-throw", function runActionForAllInSeriesThrowsTest(done) {
         var a = Defer.runActionForAllInSeries(runInSeriesArgs, function (dfd, num) { return (num > 2 ? dfd.reject(num * 2) : dfd.resolve(num * 2)); }, true).done(function (res) {
-            asr.equal(false, true, "unexpected success: " + JSON.stringify(res));
+            asr.ok(false, "unexpected success: " + JSON.stringify(res));
             done();
         }, function (err) {
             asr.equal(err, 6);